Mving smsUpdate code from MessageUtil.java class to MessageWorker.java class
